call_notes_app/agentcore_agent/research_agent/requirements.txt
bedrock-agentcore
strands-agents
boto3>=1.35.0
ddgs>=9.0.0
